+++ protected/tutorial_actions.php
@@ -159,7 +159,7 @@ for ($x=1; $x<=intval($num); $x++)
     $attrs['userPassword'] = $pw_hash;
     //    $attrs['mail'] = $email;
     $attrs['eduPersonAffiliation'][] = "member";
-    $attrs['eduPersonAffiliation'] []= "staff";
+    $attrs['eduPersonAffiliation'] []= "library-walk-in"; // Limited values are legal here. For us, this means 'tutorial'
     $attrs['telephoneNumber'] = $org_phone;
     $attrs['o'] = "BBN";
     $attrs['uidNumber'] = $id;